Implemented Sample Based Round Robin
For each service, a number of nodes(maxapps/ maxplatforms) are randomly
sampled from the population. The node with the least connections in the
sample is selected and is checked to see if it can connect with the
app. If it can't, the process is repeated. This happens a number of
times(platform minimum size)